+++ src/index.ts
@@ -18,7 +18,8 @@ import {MercatorCoordinate} from './geo/mercator_coordinate';
 import {Evented, type ErrorEvent, Event, type Listener} from './util/evented';
 import {type AddProtocolAction, config} from './util/config';
 import {rtlMainThreadPluginFactory} from './source/rtl_text_plugin_main_thread';
-import {WorkerPool} from './util/worker_pool';
+import {setNow, restoreNow, isTimeFrozen} from './util/time_control';
+import { WorkerPool } from './util/worker_pool';
 import {prewarm, clearPrewarmedResources} from './util/global_worker_pool';
 import {AJAXError, type ExpiryData, type GetResourceResponse, type RequestParameters} from './util/ajax';
 import {GeoJSONSource, type SetClusterOptions} from './source/geojson_source';
@@ -368,5 +369,8 @@ export {
     removeProtocol,
     addSourceType,
     importScriptInWorkers,
-    createTileMesh
+    createTileMesh,
+    setNow,
+    restoreNow,
+    isTimeFrozen
 };